chore(clippy): green the M6-crate lint lane, gate the fork separately
Makefile clippy now lints an explicit M6_CRATES include-list with -D warnings
plus the m6-user sysroot lane; the vendored Starnix fork moves to an
informational `clippy-fork` target (BSD-2-Clause upstream-tracking, not churned
for our lints). Fix the warnings in the gated crates: redundant u64 casts
(smmu, iommu_ops), collapsible match arms (pcie), needless range loop (main),
dead-code helpers (rt), unused feature(linkage) (m6-std), and btree_map entry
API (zx-shim vmo).

+++ m6-kernel/src/smmu/mod.rs
@@ -966,7 +966,7 @@ pub unsafe fn init(smmu_phys: u64, smmu_virt: u64, index: u8) -> Result<(), Smmu
 
     // Diagnostic: log stream table address and alignment
     m6_pal::console::puts("[SMMU] Stream table: phys=0x");
-    put_hex_u64(strtab_phys as u64);
+    put_hex_u64(strtab_phys);
     m6_pal::console::puts(" size=0x");
     put_hex_u64(strtab_size as u64);
     m6_pal::console::puts(" align=");
